gcpt: [wip] add arguments for taking/restoring gcpt
- On taking gcpt, both a restorer and an image must be provided
    and will be loaded as follows
    |0x80000000 ~ 0x800a0000-1  | 0x800a0000 ~  |
    | Restorer                  | Image         |
- On restoring gcpt, a restorer is optional. If not provided, the
    shipped restorer in the gcpt will be used to restoring arch states
- When neither restoring nor taking, a restorer should not be provided

+++ src/monitor/ui.c
@@ -36,7 +36,7 @@ static char* rl_gets() {
 }
 
 static int cmd_c(char *args) {
-  cpu_exec((args == NULL) ? -1 : atoi(args));
+  cpu_exec((args == NULL) ? -1 : (uint64_t) atol(args));
   return 0;
 }
